requirements.txt
apache-airflow
psycopg2-binary
pandas
kafka-python
pyspark
confluent-kafka